issou

+++ back/fileManager.js
@@ -4,25 +4,35 @@ const path = require('path');
 module.exports = {
 
   addEpisode(episode) {
-     fs.writeFile('episodes/' + episode.id + '.json', JSON.stringify(episode), 'utf8', null);
+
+     fs.writeFile('episodes/' + episode.id + '.json', JSON.stringify(episode,null,2), 'utf8',  (err) => {
+         if (err) return false;
+     });
+     return true;
  },
 
- getEpisodes(dirname, onFileContent, onError) {
+ getEpisodes(dirname) {
+
+      data =[];
+
      fs.readdir(dirname, function(err, filenames) {
          if (err) {
-             onError(err);
-             return;
+             console.log("erreur reading directory");
+             return false;
          }
          filenames.forEach(function(filename) {
+
              fs.readFile(dirname + filename, 'utf-8', function(err, content) {
                  if (err) {
-                     onError(err);
-                     return;
+                     return false;
                  }
-                 onFileContent(filename, content);
+                 data.push(JSON.parse(content));
+
              });
          });
      });
+     return data;
+
  },
 
  editEpisode(uuid, name, code, note) {
@@ -40,6 +50,11 @@ module.exports = {
 
  },
     removeEpisode(uuid) {
+        if(!fs.existsSync(file)) {
+            return false;
+        }
         fs.unlinkSync('episodes/' + uuid + '.json');
+
+        return true;
     }
 }
